admin guard on gateway

+++ gateway/src/auth/auth.controller.ts
@@ -10,34 +10,50 @@ import {
   Delete,
   Res,
   UseGuards,
+  Req,
+  UnauthorizedException
 } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthBodyType } from './models/auth-body.type';
-import { Observable, take, tap } from 'rxjs';
+import { Observable, of, switchMap, take, tap } from 'rxjs';
 import { TokenType } from './models/token.type';
 import { AuthDto } from './dto/create-auth.dto';
 import { Response } from 'express';
 import { UpdateAuthDto } from './dto/update-auth.dto';
 import { AdminOrSuperAdminGuard } from './guards/admin-super-admin.guard';
+import { Request } from 'express'
 
 @Controller('auth')
 export class AuthController {
+  
   constructor(private _authService: AuthService) {}
 
   //Add an account
   @UseGuards(AdminOrSuperAdminGuard)
   @Post()
-  add(@Body() auth: UpdateAuthDto): Observable<AuthDto> {
-    return this._authService.add(auth).pipe(take(1));
+  add(@Req() req: Request, @Body() auth: UpdateAuthDto): Observable<AuthDto> {
+    //Check if an admin tries to add an auth with admin or super admin role
+    //Only a super admin is allow to
+    if(req['user'].role == 'admin' && (auth.role == 'admin' || auth.role == 'super_admin')){
+      throw new UnauthorizedException("Vous n'avez pas les droits pour attribuer ce role");
+    } else return this._authService.add(auth).pipe(take(1));
   }
 
   //Update an account
   @UseGuards(AdminOrSuperAdminGuard)
   @Patch(':id')
-  update(@Param('id') id: number, @Body() auth: UpdateAuthDto): Observable<AuthDto> {
-    return this._authService.update(id, auth).pipe(
-      take(1), // autre facon d'arrêter d'observer
-    );
+  update(@Req() req: Request, @Param('id') id: number, @Body() auth: UpdateAuthDto): Observable<AuthDto> {
+    //Check if an admin tries to modify an admin or a super admin
+    this.findOne(id).pipe(take(1)).subscribe((authSub: AuthDto) => {
+      if (req['user'].role == 'admin' && (authSub.role == 'admin' || authSub.role == 'super_admin')) {
+        throw new UnauthorizedException("Vous n'avez pas les droits pour modifier cet utilisateur");
+      }
+    })
+
+    //Check if an admin tries to modify the role of an auth to admin or super admin
+    if(req['user'].role == 'admin' && (auth.role == 'admin' || auth.role == 'super_admin')){
+      throw new UnauthorizedException("Vous n'avez pas les droits pour attribuer ce role");
+    } else return this._authService.update(id, auth).pipe(take(1));
   }
 
   //Get all account
@@ -57,8 +73,18 @@ export class AuthController {
   //Delete an account
   @UseGuards(AdminOrSuperAdminGuard)
   @Delete(':id')
-  delete(@Param('id') id: number): Observable<AuthDto> {
-    return this._authService.delete(id);
+  delete(@Req() req: Request, @Param('id') id: number): Observable<AuthDto> {
+    //Check if an admin is trying to delete an admin or a super admin
+    //We only have the id of the auth, so we need to find the auth
+    //And then we can check the role. The switchMap is use to transform an observable to another
+    return this.findOne(id).pipe(take(1), switchMap((auth: AuthDto) => {
+        if (req['user'].role == 'admin' && auth.role == 'stagiaire') {
+          return this._authService.delete(id);
+        } else {
+          throw new UnauthorizedException("Vous n'avez pas les droits pour supprimer cet utilisateur");
+        }
+      })
+    );
   }
 
   //Create a token when logged, and insert it in a cookie
@@ -73,7 +99,6 @@ export class AuthController {
 
       return token;
     } catch (error) {
-      console.error('Login error:', error);
       return { 
         error: 'Login failed' 
       };
